changed put request to include id and userId when a blog is updated

+++ controllers/blogs.js
@@ -1,32 +1,16 @@
 /* eslint-disable no-undef */
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) =>{
 	const blogs = await Blog.find({}).populate('userId', { username: 1, name: 1, id: 1 })
 	response.json(blogs)
-	
 })
-
-const getTokenFrom = request => {
-	const authorization = request.get('authorization')
-	if(authorization && authorization.startsWith('Bearer ')){
-		return authorization.replace('Bearer ', '')
-	}
-	return null
-}
 
 blogsRouter.post('/', async (request, response) => {
 	const body = request.body
-
-	const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-	if(!decodedToken.id){
-		return response.status(401).json({ error: 'token invalid' })
-	}
-	const user = await User.findById(decodedToken.id)
-
+	const user = request.user
+	
 	const blog = new Blog({
 		title: body.title,
 		author: body.author,
@@ -48,6 +32,7 @@ blogsRouter.post('/', async (request, response) => {
 
 	response.status(201).json(savedBlog)
 })
+
 blogsRouter.get('/:id', async (request, response) => {
 	const blog = await Blog.findById(request.params.id)
 	if(blog)
@@ -58,8 +43,13 @@ blogsRouter.get('/:id', async (request, response) => {
 	}
 })
 blogsRouter.delete('/:id', async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id)
-	response.status(204).end()
+	const user = request.user
+	const blog = await Blog.findById(request.params.id)
+	if(user.id.toString() === blog.userId.toString())
+	{
+		await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).end()
+	}
 })
 blogsRouter.put('/:id', async (request, response) => {
 	const body = request.body
@@ -68,7 +58,9 @@ blogsRouter.put('/:id', async (request, response) => {
 		title:body.title,
 		author: body.author,
 		url: body.url,
-		likes: body.likes
+		likes: body.likes,
+		id:body.id,
+		userId:body.userId
 	}
 
 	const updatedBlog = await Blog.findByIdAndUpdate(request.params.id,blog, {new:true})
